% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week4.R
\docType{data}
\name{binom_sample}
\alias{binom_sample}
\title{Binomial Sample}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{x}{has a 30 percent chance of being a 1, success rate is thus 30 percent}
  \item{y}{has a 70 percent chance of being a 1, success rate is thus 70 percent}
}
}
\usage{
binom_sample
}
\description{
A dataset containing two columns of randomly generated binary values
}
\keyword{datasets}
