% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week4.R
\name{prop_test}
\alias{prop_test}
\title{Two Sample Proportions Test}
\usage{
prop_test(s1, n1, s2, n2)
}
\arguments{
\item{s1}{Number of success in sample 1}

\item{n1}{Sample size 1}

\item{s2}{Number of success in sample 2}

\item{n2}{Sample size 2}
}
\value{
p-value for 95 percent confidence
}
\description{
Two Sample Proportions Test
}
\examples{
prop_test(0.5, 30, 0.8, 30)
}
